requirements.txt
anyio==4.7.0
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
